pyproject.toml
[tool.poetry]
name = "ksclothing"
version = "0.1.0"
description = ""
authors = ["Anderson Ribeiro Lopes <73368853+andeen171@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
pillow = "^9.4.0"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
